pyproject.toml
[tool.poetry]
name = "User Expense"
version = "0.1.0"
description = "j"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
daphne = "^4.0.0"
html5lib = "^1.1"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
faker = "^30.1.0"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
